pyproject.toml
[project]
name = "hello-world"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Aasher Kamal", email = "aasherkamal786@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "chainlit>=2.4.1",
    "litellm>=1.63.11",
    "python-dotenv>=1.0.1",
]

